moving patch endpoint test
